Implement q_remove_head in queue.c
This function remove the element from the head of queue
and return the pointer of the removed element.

The function returns NULL by checking if the queue is NULL
or empty. Then, it find the first entry of the queue, copying
the value of element to "sp" if non-NULL, and remove the element
from the queue.

Change-Id: Ic695e46a4c7b42279f1fda3b694bb8096c77fbe9