Implement q_remove_tail in queue.c
This function remove the element from tail of queue
and return the pointer of the removed element.

The function returns NULL by checking if the queue is NULL
or empty. Then, it finds the last entry of the queue, copying
the value of element to "sp" if non-NULL, and remove the element
from the queue.

Change-Id: Ice10a595eaf2965d0534bc1d04c350413475e021