Implement q_delete_mid in queue.c
This function delete the middle node in queue.

It finds the middle node using two pointers from head
and tail of the queue, making two pointers move towards
middle until they are point to the same node or they are
neighbor to each other. This implementation can also dealt with
null, empty or singular node queue.

Change-Id: Ia7113a2493a1c5a32cc31ebb54358340c762847d