Implement q_reverseK in queue.c
Given the node of a linked list, reverse the nodes of the list
k at a time.

At the beginning, the last node of the list is recorded as the tail
of the list. Next, there is an operation to insert the first k nodes
of the list into the tail of the list in order from back to front.
Repeat this process until there are fewer than k nodes.

Change-Id: I6f3c86ff0c601f03e80c90e61bd3cbd0bd065716