Implement q_ascend q_descend in queue.c
The function is implemented in the think that we
let the list be a stack which its head is considered
to be the bottom of the stack.

We maintain a pointer node that always point to the top
of the stack, and check the point neighbor to its left if
it is less/greater than itself.

If the left node is less/greater than the current top node,
deleting it. Otherwise, keep moving the pointer forward
until it point to the head of the list or the list is empty.

Change-Id: Ieca0b8f32af61c38e568d51401249a74cc4359f6